Module Gen.Restart

type 'a t = unit -> 'a gen

type 'a restartable = 'a t

include S with type 'a t := 'a restartable

val empty : 'a restartable

Empty generator, with no elements

val singleton : 'a -> 'a restartable

One-element generator

val return : 'a -> 'a restartable

Alias to singleton

val repeat : 'a -> 'a restartable

Repeat same element endlessly

val iterate : 'a -> ('a -> 'a) -> 'a restartable

iterate x f is [x; f x; f (f x); f (f (f x)); ...]

val unfold : ('b -> ('a * 'b) option) -> 'b -> 'a restartable

Dual of fold, with a deconstructing operation. It keeps on unfolding the 'b value into a new 'b, and a 'a which is yielded, until None is returned.

val init : ?limit:int -> (int -> 'a) -> 'a restartable

Calls the function, starting from 0, on increasing indices. If limit is provided and is a positive int, iteration will stop at the limit (excluded). For instance init ~limit:4 id will yield 0, 1, 2, and 3.

Basic combinators

Note: those combinators, applied to generators (not restartable generators) consume their argument. Sometimes they consume it lazily, sometimes eagerly, but in any case once f gen has been called (with f a combinator), gen shouldn't be used anymore.

val is_empty : _ restartable -> bool

Check whether the gen is empty. Pops an element, if any

val fold : ('b -> 'a -> 'b) -> 'b -> 'a restartable -> 'b

Fold on the generator, tail-recursively. Consumes the generator.

val reduce : ('a -> 'a -> 'a) -> 'a restartable -> 'a

Fold on non-empty sequences. Consumes the generator.

raises Invalid_argument
on an empty gen

val scan : ('b -> 'a -> 'b) -> 'b -> 'a restartable -> 'b restartable

Like fold, but keeping successive values of the accumulator. Consumes the generator.

val unfold_scan : ('b -> 'a -> 'b * 'c) -> 'b -> 'a restartable -> 'c restartable

A mix of unfold and scan. The current state is combined with the current element to produce a new state, and an output value of type 'c.

val iter : ('a -> unit) -> 'a restartable -> unit

Iterate on the gen, consumes it.

val iteri : (int -> 'a -> unit) -> 'a restartable -> unit

Iterate on elements with their index in the gen, from 0, consuming it.

val length : _ restartable -> int

Length of an gen (linear time), consuming it

val map : ('a -> 'b) -> 'a restartable -> 'b restartable

Lazy map. No iteration is performed now, the function will be called when the result is traversed.

val mapi : (int -> 'a -> 'b) -> 'a restartable -> 'b restartable

Lazy map with indexing starting from 0. No iteration is performed now, the function will be called when the result is traversed.

val fold_map : ('b -> 'a -> 'b) -> 'b -> 'a restartable -> 'b restartable

Lazy fold and map. No iteration is performed now, the function will be called when the result is traversed. The result is an iterator over the successive states of the fold.

val append : 'a restartable -> 'a restartable -> 'a restartable

Append the two gens; the result contains the elements of the first, then the elements of the second gen.

val flatten : 'a Gen_intf.gen restartable -> 'a restartable

Flatten the generator of generators

val flat_map : ('a -> 'b Gen_intf.gen) -> 'a restartable -> 'b restartable

Monadic bind; each element is transformed to a sub-gen which is then iterated on, before the next element is processed, and so on.

val mem : ?eq:('a -> 'a -> bool) -> 'a -> 'a restartable -> bool

Is the given element, member of the gen?

val take : int -> 'a restartable -> 'a restartable

Take at most n elements

val drop : int -> 'a restartable -> 'a restartable

Drop n elements

val nth : int -> 'a restartable -> 'a

n-th element, or Not_found

raises Not_found
if the generator contains less than n arguments

val take_nth : int -> 'a restartable -> 'a restartable

take_nth n g returns every element of g whose index is a multiple of n. For instance take_nth 2 (1--10) |> to_list will return 1;3;5;7;9

val filter : ('a -> bool) -> 'a restartable -> 'a restartable

Filter out elements that do not satisfy the predicate.

val take_while : ('a -> bool) -> 'a restartable -> 'a restartable

Take elements while they satisfy the predicate. The initial generator itself is not to be used anymore after this.

val fold_while : ('a -> 'b -> 'a * [ `Stop | `Continue ]) -> 'a -> 'b restartable -> 'a

Fold elements until ('a, `Stop) is indicated by the accumulator.

val drop_while : ('a -> bool) -> 'a restartable -> 'a restartable

Drop elements while they satisfy the predicate. The initial generator itself should not be used anymore, only the result of drop_while.

val filter_map : ('a -> 'b option) -> 'a restartable -> 'b restartable

Maps some elements to 'b, drop the other ones

val zip_index : 'a restartable -> (int * 'a) restartable

Zip elements with their index in the gen

val unzip : ('a * 'b) restartable -> 'a restartable * 'b restartable

Unzip into two sequences, splitting each pair

val partition : ('a -> bool) -> 'a restartable -> 'a restartable * 'a restartable

partition p l returns the elements that satisfy p, and the elements that do not satisfy p

val for_all : ('a -> bool) -> 'a restartable -> bool

Is the predicate true for all elements?

val exists : ('a -> bool) -> 'a restartable -> bool

Is the predicate true for at least one element?

val min : ?lt:('a -> 'a -> bool) -> 'a restartable -> 'a

Minimum element, according to the given comparison function.

raises Invalid_argument
if the generator is empty

val max : ?lt:('a -> 'a -> bool) -> 'a restartable -> 'a

Maximum element, see min

raises Invalid_argument
if the generator is empty

val eq : ?eq:('a -> 'a -> bool) -> 'a restartable -> 'a restartable -> bool

Equality of generators.

val lexico : ?cmp:('a -> 'a -> int) -> 'a restartable -> 'a restartable -> int

Lexicographic comparison of generators. If a generator is a prefix of the other one, it is considered smaller.

val compare : ?cmp:('a -> 'a -> int) -> 'a restartable -> 'a restartable -> int

Synonym for lexico

val find : ('a -> bool) -> 'a restartable -> 'a option

find p e returns the first element of e to satisfy p, or None.

val sum : int restartable -> int

Sum of all elements

Multiple iterators

val map2 : ('a -> 'b -> 'c) -> 'a restartable -> 'b restartable -> 'c restartable

Map on the two sequences. Stops once one of them is exhausted.

val iter2 : ('a -> 'b -> unit) -> 'a restartable -> 'b restartable -> unit

Iterate on the two sequences. Stops once one of them is exhausted.

val fold2 : ('acc -> 'a -> 'b -> 'acc) -> 'acc -> 'a restartable -> 'b restartable -> 'acc

Fold the common prefix of the two iterators

val for_all2 : ('a -> 'b -> bool) -> 'a restartable -> 'b restartable -> bool

Succeeds if all pairs of elements satisfy the predicate. Ignores elements of an iterator if the other runs dry.

val exists2 : ('a -> 'b -> bool) -> 'a restartable -> 'b restartable -> bool

Succeeds if some pair of elements satisfy the predicate. Ignores elements of an iterator if the other runs dry.

val zip_with : ('a -> 'b -> 'c) -> 'a restartable -> 'b restartable -> 'c restartable

Combine common part of the gens (stops when one is exhausted)

val zip : 'a restartable -> 'b restartable -> ('a * 'b) restartable

Zip together the common part of the gens

Complex combinators

val merge : 'a Gen_intf.gen restartable -> 'a restartable

Pick elements fairly in each sub-generator. The merge of gens e1, e2, ... picks elements in e1, e2, in e3, e1, e2 .... Once a generator is empty, it is skipped; when they are all empty, and none remains in the input, their merge is also empty. For instance, merge [1;3;5] [2;4;6] will be, in disorder, 1;2;3;4;5;6.

val intersection : ?cmp:('a -> 'a -> int) -> 'a restartable -> 'a restartable -> 'a restartable

Intersection of two sorted sequences. Only elements that occur in both inputs appear in the output

val sorted_merge : ?cmp:('a -> 'a -> int) -> 'a restartable -> 'a restartable -> 'a restartable

Merge two sorted sequences into a sorted sequence

val sorted_merge_n : ?cmp:('a -> 'a -> int) -> 'a restartable list -> 'a restartable

Sorted merge of multiple sorted sequences

val tee : ?n:int -> 'a restartable -> 'a Gen_intf.gen list

Duplicate the gen into n generators (default 2). The generators share the same underlying instance of the gen, so the optimal case is when they are consumed evenly

val round_robin : ?n:int -> 'a restartable -> 'a Gen_intf.gen list

Split the gen into n generators in a fair way. Elements with index = k mod n with go to the k-th gen. n default value is 2.

val interleave : 'a restartable -> 'a restartable -> 'a restartable

interleave a b yields an element of a, then an element of b, and so on. When a generator is exhausted, this behaves like the other generator.

val intersperse : 'a -> 'a restartable -> 'a restartable

Put the separator element between all elements of the given gen

val product : 'a restartable -> 'b restartable -> ('a * 'b) restartable

Cartesian product, in no predictable order. Works even if some of the arguments are infinite.

val group : ?eq:('a -> 'a -> bool) -> 'a restartable -> 'a list restartable

Group equal consecutive elements together.

val uniq : ?eq:('a -> 'a -> bool) -> 'a restartable -> 'a restartable

Remove consecutive duplicate elements. Basically this is like fun e -> map List.hd (group e).

val sort : ?cmp:('a -> 'a -> int) -> 'a restartable -> 'a restartable

Sort according to the given comparison function. The gen must be finite.

val sort_uniq : ?cmp:('a -> 'a -> int) -> 'a restartable -> 'a restartable

Sort and remove duplicates. The gen must be finite.

val chunks : int -> 'a restartable -> 'a array restartable

chunks n e returns a generator of arrays of length n, composed of successive elements of e. The last array may be smaller than n

val permutations : 'a restartable -> 'a list restartable

Permutations of the gen.

val permutations_heap : 'a restartable -> 'a array restartable

Permutations of the gen, using Heap's algorithm.

val combinations : int -> 'a restartable -> 'a list restartable

Combinations of given length. The ordering of the elements within each combination is unspecified. Example (ignoring ordering): combinations 2 (1--3) |> to_list = [[1;2]; [1;3]; [2;3]]

val power_set : 'a restartable -> 'a list restartable

All subsets of the gen (in no particular order). The ordering of the elements within each subset is unspecified.

Basic conversion functions

val of_list : 'a list -> 'a restartable

Enumerate elements of the list

val to_list : 'a restartable -> 'a list

non tail-call trasnformation to list, in the same order

val to_rev_list : 'a restartable -> 'a list

Tail call conversion to list, in reverse order (more efficient)

val to_array : 'a restartable -> 'a array

Convert the gen to an array (not very efficient)

val of_array : ?start:int -> ?len:int -> 'a array -> 'a restartable

Iterate on (a slice of) the given array

val of_string : ?start:int -> ?len:int -> string -> char restartable

Iterate on bytes of the string

val to_string : char restartable -> string

Convert into a string

val to_buffer : Buffer.t -> char restartable -> unit

Consumes the iterator and writes to the buffer

val rand_int : int -> int restartable

Random ints in the given range.

val int_range : ?step:int -> int -> int -> int restartable

int_range ~step a b generates integers between a and b, included, with steps of length step (1 if omitted). a is assumed to be smaller than b, otherwise the result will be empty.

raises Invalid_argument
if step=0

parameter step
step between two numbers; must not be zero, but it can be negative for decreasing ranges (@since 0.5).

val lines : char restartable -> string restartable

Group together chars belonging to the same line

val unlines : string restartable -> char restartable

Explode lines into their chars, adding a '\n' after each one

module Infix : sig ... end

val (--) : int -> int -> int restartable

Synonym for int_range ~by:1

val (>>=) : 'a restartable -> ('a -> 'b Gen_intf.gen) -> 'b restartable

Monadic bind operator

val (>>|) : 'a restartable -> ('a -> 'b) -> 'b restartable

Infix map operator

val (>|=) : 'a restartable -> ('a -> 'b) -> 'b restartable

Infix map operator

val pp : ?start:string -> ?stop:string -> ?sep:string -> ?horizontal:bool -> (Format.formatter -> 'a -> unit) -> Format.formatter -> 'a restartable -> unit

Pretty print the content of the generator on a formatter.

val of_seq : 'a Seq.t -> 'a restartable

val to_iter : 'a restartable -> 'a Gen_intf.iter

val cycle : 'a t -> 'a t

Cycle through the enum, endlessly. The enum must not be empty.

val lift : ('a gen -> 'b) -> 'a t -> 'b

val lift2 : ('a gen -> 'b gen -> 'c) -> 'a t -> 'b t -> 'c

val of_gen : ?caching:bool -> ?max_chunk_size:int -> 'a gen -> 'a t

Use persistent_lazy to convert a one-shot generator into a restartable one. See GenMList.of_gen_lazy for more details on parameters.