Module GenLabels.IO

val with_in : ?mode:int -> ?flags:open_flag list -> file:string -> (char t -> 'a) -> 'a

with_in ~file f opens file and calls f g, where g is a generator of characters from the file. The generator is only valid within the scope in which f is called.

val with_lines : ?mode:int -> ?flags:open_flag list -> file:string -> (string t -> 'a) -> 'a

with_lines ~file f opens file file and calls f g, where g is a generator that iterates on the lines from the file. Do not use the generator outside of the scope of f

val write_str : ?mode:int -> ?flags:open_flag list -> ?sep:string -> file:string -> string t -> unit

write_to ~file g writes all strings from g into the given file. It takes care of opening and closing the file. Does not add sep after the last string.

parameter mode
default 0o644

parameter flags
used by open_out_gen. Default: [Open_creat;Open_wronly].

parameter sep
separator between each string (e.g. newline)

val write : ?mode:int -> ?flags:open_flag list -> file:string -> char t -> unit

Same as write_str but with individual characters

val write_lines : ?mode:int -> ?flags:open_flag list -> file:string -> string t -> unit

write_lines ~file g is similar to write_str file g ~sep:"\n" but also adds '\n' at the end of the file